+++ src/lib.rs
@@ -31,6 +31,6 @@ mod templates;
 
 pub use site::{Site};
 pub use config::{Config, get_config};
-pub use front_matter::{PageFrontMatter, SectionFrontMatter, split_page_content, split_section_content};
+pub use front_matter::{PageFrontMatter, SectionFrontMatter, InsertAnchor, split_page_content, split_section_content};
 pub use content::{Page, Section, SortBy, sort_pages, populate_previous_and_next_pages};
 pub use fs::{create_file};